Splitting out List.prototype.underscoreMethod, list.prototype.getDocumentName
