+++ lib/list.js
@@ -116,8 +116,10 @@ function List(key, options) {
 
 // Add prototype methods
 List.prototype.relationship = require('./list/relationship');
+List.prototype.underscoreMethod = require('./list/underscoreMethod');
 List.prototype.getData = require('./list/getData');
 List.prototype.getOptions = require('./list/getOptions');
+List.prototype.getDocumentName = require('./list/getDocumentName');
 List.prototype.addFiltersToQuery = require('./list/addFiltersToQuery');
 List.prototype.isReserved = require('./list/isReserved');
 List.prototype.expandColumns = require('./list/expandColumns');
@@ -295,21 +297,7 @@ List.prototype.field = function(path, options) {
 };
 
 
-/**
- * Adds a method to the underscoreMethods collection on the list, which is then
- * added to the schema before the list is registered with mongoose.
- */
-List.prototype.underscoreMethod = function(path, fn) {
-	var target = this.underscoreMethods;
-	path = path.split('.');
-	var last = path.pop();
-	path.forEach(function(part) {
-		if (!target[part]) target[part] = {};
-		target = target[part];
-	});
-	target[last] = fn;
-	return this;
-};
+
 
 
 /**
@@ -374,21 +362,6 @@ List.prototype.register = function() {
 	}
 	require('../').list(this);
 	return this;
-};
-
-
-/**
- * Gets the name of the provided document from the correct path
- *
- * Example:
- *     var name = list.getDocumentName(item)
- *
- * @param {Object} item
- * @param {Boolean} escape - causes HTML entities to be encoded
- */
-List.prototype.getDocumentName = function(doc, escape) {
-	var name = String(this.nameField ? this.nameField.format(doc) : doc.get(this.namePath));
-	return (escape) ? utils.encodeHTMLEntities(name) : name;
 };
